requirements.txt
playwright
python-dotenv
google-genai
Pillow
requests
